# perf(gatsby): pull in cache lib and change lock from fs to mem