… packages/gatsby/package.json
… packages/gatsby/src/utils/cache.ts
… yarn.lock