The cache object that we send to plugins relies on a fs-based locking mechanism to prevent race conditions from clobbering cache files. However this is causing a lot of io pressure, to the point where a baseline markdown benchmark drops 10%-15% just for this change.

Markdown is significantly affected because it aggressively caches the parsing steps. Other plugins might too, I've seen improvements in MDX for example. But really anything that heavily calls `cache.get` and `cache.set` should see a boost from this change.

This PR pulls in the plugin that we use for this cache: https://github.com/rolandstarke/node-cache-manager-fs-hash
The first commit copies it and only applies prettier/lint style fixes.
The second commit makes TS happy
The third commit replaces the `lockfile` library with a global map with the same purpose.

The in-memory lock includes a 400ms debounce and 10s timeout, just in case.

From my brief testing, I'm not sure we need the lock at all but for now it's probably better to keep it. There appears to be no significant difference in keeping the in-memory version anyways so that's fine.

Some numbers on markdown and mdx. I didn't run the full mdx benchmark because it literally takes over an hour. I think the impact should be similar to markdown in absolute terms (not relative) because mdx has a lot of other pain points that slow it down. Stats will tell in due time.

```
(bs=bootstrap, rq=run queries)
gabe-fs-markdown
    fs mem lock:
        - 10k:   40s (10s bs,  21s rq)
        - 100k: 372s (77s bs, 254s rq)
    mem lock:
        - 10k:   37s (10s bs,  18s rq)
        - 100k: 336s (78s bs, 217s rq)
    no lock:
        - 10k:   37s (10s bs,  18s rq)
        - 100k: 334s (78s bs, 215s rq)
gabe-fs-mdx
    fs lock:
        - 10k:  304s (117s bs, 177s rq)
    mem lock:
        - 10k:  297s (115s bs, 171s rq)
    no lock:
        - 10k:  295s (114s bs, 171s rq)
```