@pvdz 2020-11-10
Awkward, good point. This is the core commit for this PR: https://github.com/gatsbyjs/gatsby/pull/27873/commits/cb2820c3e4191addd937b957e5692f99c25ff095

I've added a commit that actually sets the lock key.